@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class RolifyCreateRoles < ActiveRecord::Migration
+class RolifyCreateRoles < ActiveRecord::Migration[5.0]
   def change
     create_table :roles, id: :serial, force: true do |t|
       t.string :name
