@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateLeads < ActiveRecord::Migration
+class CreateLeads < ActiveRecord::Migration[5.0]
   def change
     create_table :leads, id: :serial, force: true do |t|
       t.string :name
