@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class NewUserFields < ActiveRecord::Migration
+class NewUserFields < ActiveRecord::Migration[5.0]
   def change
     # add_column :users, :state,	:string
     # add_column :users, :branch,	:string
