@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class RemoveDeviseNullAttrib < ActiveRecord::Migration
+class RemoveDeviseNullAttrib < ActiveRecord::Migration[5.0]
   def change
     change_column :users, :email, :string, null: true
     change_column :users, :encrypted_password, :string, null: true
