@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateRosters < ActiveRecord::Migration
+class CreateRosters < ActiveRecord::Migration[5.0]
   def change
     create_table :rosters do |t|
       t.string :organisation
