@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreatePatrolMembers < ActiveRecord::Migration
+class CreatePatrolMembers < ActiveRecord::Migration[5.0]
   def change
     create_table :patrol_members do |t|
       t.integer :user_id
