@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateRequests < ActiveRecord::Migration
+class CreateRequests < ActiveRecord::Migration[5.0]
   def change
     create_table :requests do |t|
       t.integer :roster_id
