@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class AddSessionsTable < ActiveRecord::Migration
+class AddSessionsTable < ActiveRecord::Migration[5.0]
   def change
     create_table :sessions do |t|
       t.string :session_id, null: false
