@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateEmails < ActiveRecord::Migration
+class CreateEmails < ActiveRecord::Migration[5.0]
   def change
     create_table :emails, &:timestamps
   end
