@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateSettings < ActiveRecord::Migration
+class CreateSettings < ActiveRecord::Migration[5.0]
   def change
     create_table :settings do |t|
       t.string :key
