@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateClubs < ActiveRecord::Migration
+class CreateClubs < ActiveRecord::Migration[5.0]
   def change
     create_table :clubs do |t|
       t.string :name
