@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateNoticeAcknowledgements < ActiveRecord::Migration
+class CreateNoticeAcknowledgements < ActiveRecord::Migration[5.0]
   def change
     create_table :notice_acknowledgements do |t|
       t.integer :user_id
