@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class CreateNotices < ActiveRecord::Migration
+class CreateNotices < ActiveRecord::Migration[5.0]
   def change
     create_table :notices do |t|
       t.string :title
