@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class AddPhoneToLeads < ActiveRecord::Migration
+class AddPhoneToLeads < ActiveRecord::Migration[5.0]
   def change
     add_column :leads, :phone, :string
   end
