@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class AddColumnsToClub < ActiveRecord::Migration
+class AddColumnsToClub < ActiveRecord::Migration[5.0]
   def change
     add_column :clubs, :lat, :float, default: 0, null: false		# Latitude for weather widget
     add_column :clubs, :lon, :float, default: 0, null: false		# Longitude for weather widget
