@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 # Migration responsible for creating a table with activities
-class CreateActivities < ActiveRecord::Migration
+class CreateActivities < ActiveRecord::Migration[5.0]
   # Create table
   def self.up
     create_table :activities do |t|
