@@ -11,7 +11,7 @@ port        ENV['PORT']     || 3000
 environment ENV['RACK_ENV'] || 'development'
 
 on_worker_boot do
-  # Worker specific setup for Rails 4.1+
+  # Worker specific setup for Rails 4.1+ but no longer needed for Rails 5.2+ apps
   # See: https://devcenter.heroku.com/articles/deploying-rails-applications-with-the-puma-web-server#on-worker-boot
   ActiveRecord::Base.establish_connection
 end
