@@ -1,9 +1,14 @@
 # frozen_string_literal: true
 
-# This file is copied to spec/ when you run 'rails generate rspec:install'
 require 'spec_helper'
+require 'public_activity/testing'
+
 ENV['RAILS_ENV'] ||= 'test'
 require File.expand_path('../config/environment', __dir__)
+
+# Don't write activity when running tests
+PublicActivity.enabled = false
+
 # Prevent database truncation if the environment is production
 abort('The Rails environment is running in production mode!') if Rails.env.production?
 require 'rspec/rails'
@@ -38,11 +43,11 @@ RSpec.configure do |config|
   # instead of true.
   config.use_transactional_fixtures = true
 
-  Capybara.default_driver = :selenium # :selenium_chrome and :selenium_chrome_headless are also registered
-
-  Capybara.current_driver = :selenium
-
-  Capybara.server = :puma # Until your setup is working
+  Capybara.default_driver = :selenium_chrome # :selenium_chrome and :selenium_chrome_headless are also registered
+  Capybara.current_driver = :selenium_chrome
+  Capybara.javascript_driver = :selenium_chrome
+  Capybara.server = :puma
+  Capybara.test_id = 'data-test-id'
 
   # Capybara.run_server = false
   # Capybara.app_host = "http://localhost:3000"
